requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.1
bidict==0.23.1
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.20
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.15
deprecation==2.1.0
distlib==0.3.9
distro==1.9.0
durationpy==0.9
fastapi==0.115.5
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
flatbuffers==24.3.25
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.8.0
google-crc32c==1.6.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gotrue==2.10.0
greenlet==3.1.1
grpcio==1.68.0
grpcio-status==1.68.0
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.1
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
kubernetes==31.0.0
langchain==0.3.7
langchain-chroma==0.1.4
langchain-community==0.3.3
langchain-core==0.3.19
langchain-google-genai==2.0.1
langchain-huggingface==0.1.2
langchain-openai==0.2.8
langchain-text-splitters==0.3.2
langsmith==0.1.146
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.14.2
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.54.4
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
orjson==3.10.12
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pip-tools==7.4.1
pipdeptree==2.24.0
platformdirs==4.3.6
postgrest==0.18.0
posthog==3.7.3
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
realtime==2.0.6
referencing==0.35.1
regex==2024.11.6
requests==2.32.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.3
rpds-py==0.21.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.1
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
storage3==0.9.0
streamlit==1.40.0
supabase==2.10.0
supafunc==0.7.0
sympy==1.13.3
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.4
toml==0.10.2
torch==2.2.2
tornado==6.4.2
tqdm==4.67.1
transformers==4.46.0
typer==0.13.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
umap==0.1.1
umap-learn==0.5.7
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
virtualenv==20.27.0
watchfiles==1.0.0
websocket-client==1.8.0
websockets==13.0
Werkzeug==3.1.3
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.0
zipp==3.21.0
